standalone ElapsedTime
